App Cleanup (#83)
The changes:
- Modal
  - Moved Modal component into it's own dir and separated components into their own files.
  - Added the ability to close modals by pressing the Escape key.
- Navigation
  - Major rework to move navigation out of pages and make it global in App.js.
  - Moved Nav files into new dir src/components/navigation and separated components into their own files.
  - Added logic to handle edge case navigation layouts such as transparent themes and custom branding.
- Dev Mode
  - Added a development banner that displays when in a development environment.
  - This is controlled by a feature flag: REACT_APP_DEVELOPMENT_BANNER
- Misc
  - Other React specific changes to improve render times as well as following typical semantics and best practices when conditionally rendering components. Most importantly, not immediately returning null when a component uses state or effects afterwards.
  - Bugfix: Allow dropdown items to close the dropdown if they are <a> tags
