+++ .github/workflows/develop.yml
@@ -17,6 +17,7 @@ jobs:
       REACT_APP_FORMULA_EDITOR: true
       REACT_APP_INSTRUMENT_CHART: true
       REACT_APP_CROSS_SECTION: true
+      REACT_APP_DEVELOPMENT_BANNER: true
     runs-on: ubuntu-latest
     strategy:
       matrix:
