+++ src/app-pages/instrument/timeseries/timeseries-list-item.js
@@ -11,34 +11,36 @@ export default connect(
   'doModalOpen',
   ({ projectsByRoute: project, doModalOpen, item, onClick, active }) => {
     const li = useRef(null);
-    if (!item) return null;
     const itemClass = classnames({
       'list-group-item': true,
       active: active,
     });
+
     return (
-      <li
-        ref={li}
-        className={itemClass}
-        onClick={(e) => {
-          if (e.currentTarget === li.current) onClick(item);
-        }}
-      >
-        <RoleFilter allowRoles={[`${project.slug.toUpperCase()}.*`]}>
-          <button
-            className='float-right btn btn-sm btn-outline-info'
-            onClick={() => {
-              doModalOpen(TimeseriesForm, { item: item });
-            }}
-          >
-            <Icon icon='pencil' />
-          </button>
-        </RoleFilter>
-        <div>{item.name}</div>
-        <div>
-          <small>{`${item.parameter} in ${item.unit}`}</small>
-        </div>
-      </li>
+      item && (
+        <li
+          ref={li}
+          className={itemClass}
+          onClick={(e) => {
+            if (e.currentTarget === li.current) onClick(item);
+          }}
+        >
+          <RoleFilter allowRoles={[`${project.slug.toUpperCase()}.*`]}>
+            <button
+              className='float-right btn btn-sm btn-outline-info'
+              onClick={() => {
+                doModalOpen(TimeseriesForm, { item: item });
+              }}
+            >
+              <Icon icon='pencil' />
+            </button>
+          </RoleFilter>
+          <div>{item.name}</div>
+          <div>
+            <small>{`${item.parameter} in ${item.unit}`}</small>
+          </div>
+        </li>
+      )
     );
   }
 );
