+++ src/app-pages/reporting/batch-plotting/batch-plotting.js
@@ -6,7 +6,6 @@ import Card from '../../../app-components/card';
 import DataConfiguration from './data-configuration';
 import Icon from '../../../app-components/icon';
 import Map from '../../../app-components/classMap';
-import Navbar from '../../../app-components/navbar';
 
 import '../reporting.scss';
 
@@ -23,7 +22,6 @@ const BatchPlotting = connect(
 
     return (
       <>
-        <Navbar theme='primary' fixed />
         <section className='container-fluid page-body'>
           <Card className='w-100'>
             <Card.Header text='Plot Data Configuration' />
