+++ src/app-pages/reporting/reporting.scss
@@ -1,5 +1,4 @@
 .page-body {
-  margin-top: 6rem;
   position: relative;
 }
